Refactor Omnipose code into omnipose/omnipose.py.
