+++ cellpose/__init__.py
@@ -1,3 +1,4 @@
 from .io import logger_setup
+import omnipose
 
 logger, log_file = logger_setup()
