+++ omnipose/README.md
@@ -1,4 +1,4 @@
 <img src="logo.png" width="150" title="omnipose" alt="omnipose" align="center" vspace = "0">
 
-The code associated with Omnipose is licensed under the
+The code associated with Omnipose is dual licensed under the
 [GNU Affero General Public License](https://www.gnu.org/licenses/agpl-3.0.en.html). For commercial use, contact uwcomotion@uw.edu.
